refactor authz middleware
