+++ internal/server/api.go
@@ -160,3 +160,14 @@ func (s *Server) apiHandleLogin(w http.ResponseWriter, r *http.Request) error {
 
 	return json.NewEncoder(w).Encode(token)
 }
+
+func (s *Server) apiHandleMe(w http.ResponseWriter, r *http.Request) error {
+	const q = "SELECT id, username, email, created_at FROM users WHERE id = $1"
+	row, err := s.dbClient.QueryRow(r.Context(), q, r.Context().Value(middleware.CtxUserKey))
+	if err != nil {
+		return err
+	}
+	var user User
+	row.Scan(&user.ID, &user.Username, &user.Email, &user.CreatedAt)
+	return json.NewEncoder(w).Encode(user)
+}
