+++ internal/server/middleware/authentication.go
@@ -9,8 +9,6 @@ import (
 	"errors"
 	"strings"
 	"time"
-
-	"golang.org/x/crypto/bcrypt"
 )
 
 type joseHeader struct {
@@ -33,8 +31,9 @@ type JWT struct {
 
 // GenerateJWT takes a user id and the signing secret and generates a JWT.
 // with the following structure:
-//  Header: {"alg":"HS256", "typ":"JWT"}
-//  Claims: {"user_id": user_id, "exp": [current time + 12hrs]}
+//
+//	Header: {"alg":"HS256", "typ":"JWT"}
+//	Claims: {"user_id": user_id, "exp": [current time + 12hrs]}
 func GenerateJWT(userID int, secret []byte) (JWT, error) {
 	// Set the header and payload
 	jwt := JWT{
@@ -72,9 +71,9 @@ func GenerateJWT(userID int, secret []byte) (JWT, error) {
 // with the fields filled out appropriately.
 func ParseJWT(s string) (JWT, error) {
 	jwtParts := strings.Split(s, ".")
-  if len(jwtParts) != 3 {
-    return JWT{}, errors.New("malformed JWT")
-  }
+	if len(jwtParts) != 3 {
+		return JWT{}, errors.New("malformed JWT")
+	}
 	headerBytes, err := base64.RawURLEncoding.DecodeString(jwtParts[0])
 	if err != nil {
 		return JWT{}, err
@@ -125,25 +124,4 @@ func (j *JWT) IsValid(secret []byte) (bool, error) {
 		return true, nil
 	}
 	return false, nil
-}
-
-// GeneratePasswordHash adds a hashed password to a User struct if  there is a
-// password in the struct, and a password hash is not already present.
-func (u *User) GeneratePasswordHash() error {
-	// If no password is provided, then it should be set to an empty string.
-	// It shouldn't be possible to set an empty string as a password in any case.
-	if u.PasswordHash != nil {
-		return errors.New("user struct already contains password hash")
-	}
-	if u.Password == "" {
-		return errors.New("attempted to generate password hash where password is an empty string")
-	}
-
-	hash, err := bcrypt.GenerateFromPassword([]byte(u.Password), bcrypt.DefaultCost)
-	if err != nil {
-		return err
-	}
-	u.PasswordHash = hash
-	u.Password = ""
-	return nil
 }
