+++ internal/server/middleware/authorization.go
@@ -1,35 +1,28 @@
 package middleware
 
 import (
+	"errors"
 	"net/http"
-
-	"github.com/jessesomerville/yodahunters/internal/log"
 )
 
-func isAuthenticated(r *http.Request, secret []byte) bool {
+// Authorize takes a request, verifies that it contains a valid
+// JWT, then returns the user id for that the user in the JWT.
+func Authorize(r *http.Request, secret []byte) (int, error) {
 	accessToken, err := r.Cookie("access_token")
 	if err != nil {
-		log.Errorf(r.Context(), "failed to read access_token cookie")
-		return false
+		return -1, err
 	}
 	jwt, err := ParseJWT(accessToken.Value)
 	if err != nil {
-		log.Errorf(r.Context(), "failed to parse jwt in access_token cookie")
-		return false
+		return -1, err
 	}
 	valid, err := jwt.IsValid(secret)
 	if err != nil {
-		log.Errorf(r.Context(), "unable to validate JWT")
-		return false
+		return -1, err
 	}
-	return valid
-}
+	if !valid {
+		return -1, errors.New("invalid JWT")
+	}
 
-func AuthorizationHandler(next http.HandlerFunc, secret []byte) http.HandlerFunc {
-	return func(w http.ResponseWriter, r *http.Request) {
-		if !isAuthenticated(r, secret) {
-			http.Redirect(w, r, "/login", http.StatusFound)
-		}
-		next.ServeHTTP(w, r)
-	}
+	return jwt.Payload.UserID, nil
 }
