+++ internal/server/models.go
@@ -2,7 +2,10 @@ package server
 
 // TODO.
 import (
+	"errors"
 	"time"
+
+	"golang.org/x/crypto/bcrypt"
 )
 
 // User is a struct for managing users in the app.
@@ -24,4 +27,25 @@ type Thread struct {
 	Title     string    `json:"Title" db:"title"`
 	Body      string    `json:"Body" db:"body"`
 	CreatedAt time.Time `json:"created_at,omitempty" db:"created_at"`
+}
+
+// GeneratePasswordHash adds a hashed password to a User struct if  there is a
+// password in the struct, and a password hash is not already present.
+func (u *User) GeneratePasswordHash() error {
+	// If no password is provided, then it should be set to an empty string.
+	// It shouldn't be possible to set an empty string as a password in any case.
+	if u.PasswordHash != nil {
+		return errors.New("user struct already contains password hash")
+	}
+	if u.Password == "" {
+		return errors.New("attempted to generate password hash where password is an empty string")
+	}
+
+	hash, err := bcrypt.GenerateFromPassword([]byte(u.Password), bcrypt.DefaultCost)
+	if err != nil {
+		return err
+	}
+	u.PasswordHash = hash
+	u.Password = ""
+	return nil
 }
